test(rust): execute logging / tracing tests as integration tests
Those tests need to be isolated since they set some global exporters
