+++ implementations/rust/ockam/ockam_api/src/cli_state/journeys/journeys.rs
@@ -162,82 +162,3 @@ impl CliState {
         (opentelemetry_context, now)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::logs::{LoggingConfiguration, LoggingTracing};
-    use crate::random_name;
-    use ockam_node::Executor;
-    use opentelemetry::trace::FutureExt;
-    use opentelemetry_sdk::testing::logs::InMemoryLogsExporter;
-    use opentelemetry_sdk::testing::trace::InMemorySpanExporter;
-    use tempfile::NamedTempFile;
-
-    #[test]
-    fn test_create_journey_event() {
-        let spans_exporter = InMemorySpanExporter::default();
-        let logs_exporter = InMemoryLogsExporter::default();
-
-        let tracing_guard = LoggingTracing::setup_with_exporters(
-            spans_exporter.clone(),
-            logs_exporter.clone(),
-            None,
-            LoggingConfiguration::off().set_crates(&["ockam_api"]),
-            "test",
-        );
-        let tracer = global::tracer("ockam-test");
-        let result = tracer.in_span("user event", |cx| {
-            let _guard = cx.with_value(Utc::now()).attach();
-
-            Executor::execute_future(
-                async move {
-                    let db_file = NamedTempFile::new().unwrap();
-                    let cli_state_directory = db_file.path().parent().unwrap().join(random_name());
-                    let cli = CliState::create(cli_state_directory)
-                        .await
-                        .unwrap()
-                        .set_tracing_enabled();
-
-                    let mut map = HashMap::new();
-                    map.insert(USER_EMAIL, "etorreborre@yahoo.com");
-                    map.insert(USER_NAME, "eric");
-                    cli.add_journey_event(JourneyEvent::Enrolled, map.clone())
-                        .await
-                        .unwrap();
-                    cli.add_journey_event(JourneyEvent::PortalCreated, map)
-                        .await
-                        .unwrap();
-                }
-                .with_current_context(),
-            )
-        });
-        assert!(result.is_ok());
-
-        tracing_guard.force_flush();
-        let mut spans = spans_exporter.get_finished_spans().unwrap();
-        spans.sort_by_key(|s| s.start_time);
-        assert_eq!(spans.len(), 4);
-
-        let span_names = spans.iter().map(|s| s.name.as_ref()).collect::<Vec<&str>>();
-        assert_eq!(
-            span_names,
-            vec![
-                "user event",
-                "start host journey",
-                "enrolled",
-                "portal created"
-            ]
-        );
-        // remove the first event
-        spans.remove(0);
-
-        // all user events have the same start/end times and have a duration of 1ms
-        let first_span = spans.first().unwrap().clone();
-        for span in spans {
-            assert_eq!(span.start_time, first_span.start_time);
-            assert_eq!(span.end_time, first_span.end_time);
-            assert_eq!(span.start_time.add(Duration::from_millis(1)), span.end_time);
-        }
-    }
-}
